+++ src/lib/core/ace/models.hpp
@@ -202,6 +202,7 @@ namespace GATAS
         FANET = 3,
         OGN1 = 4,
         PAW = 5,
+        _RADIO = 6,          // ANything beflore Radio can only be received over hardware
         ADSB = 6,
         _TRANSPROTOCOLS = 6, // Indicate maximum RADIO that can be received over low power (868MHZ etc..) used to limit array sizes
         ADSLFLARM = 253,     // Combination of ADSL/FLARM, not an acutal protocol but needed for RX of multiple protocols